+++ web/modules/custom/ai_screening_project_track/src/Theme/ThemeNegotiator.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Drupal\ai_screening_project_track\Theme;
+
+use Drupal\Core\Routing\RouteMatchInterface;
+use Drupal\Core\Theme\ThemeNegotiatorInterface;
+
+/**
+ * Provides a ThemeNegotiator for content type gist.
+ */
+class ThemeNegotiator implements ThemeNegotiatorInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function applies(RouteMatchInterface $route_match): bool {
+    return NULL !== $route_match->getParameter('webform_submission');
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function determineActiveTheme(RouteMatchInterface $route_match) {
+    return 'itkdev_project_theme';
+  }
+
+}